preview room now available when user tap on a room
