+++ DesignScape/DesignScape/Controller/UserManager.swift
@@ -123,4 +123,28 @@ extension UserManager {
             }
         }
     }
+    
+    func downloadRoomAsync(fileRef: StorageReference) async throws -> URL {
+        return try await withCheckedThrowingContinuation { continuation in
+            downloadRoom(fileRef: fileRef) { url in
+                continuation.resume(returning: url)
+            }
+        }
+    }
+    
+    func downloadRoom(fileRef: StorageReference, onSuccess: @escaping ((URL) -> Void)) {
+        // Create a temporary file URL
+        let tempDirectoryURL = FileManager.default.temporaryDirectory
+        let tempFileURL = tempDirectoryURL.appendingPathComponent(UUID().uuidString).appendingPathExtension("usdz")
+        
+        // Download the USDZ file to the temporary file URL
+        fileRef.write(toFile: tempFileURL) { result in
+            switch result {
+            case .success(let url):
+                onSuccess(url)
+            case .failure(_):
+                print("Error downloading USDZ")
+            }
+        }
+    }
 }
