+++ DesignScape/DesignScape/View/MainView.swift
@@ -22,7 +22,7 @@ struct MainView: View {
                         selectedTab == 1 ? Image("home.fill") : Image("home")
                     }
                     .tag(1)
-                RoomLoaderView()
+                UserRoomsView()
                     .tabItem {
                         Text("Copilot")
                         selectedTab == 2 ? Image("stars.fill") : Image("stars")
