+++ DesignScape/DesignScape/View/Profile/UserRoomsView.swift
@@ -22,29 +22,31 @@ struct UserRoomsView: View {
                     )
                 Spacer()
             }
-            LazyVGrid(columns: [GridItem(.flexible(), spacing: 16),
-                                GridItem(.flexible(), spacing: 16)], spacing: 16) {
-                ForEach(usdzFiles, id: \.self) { fileRef in
-                    NavigationLink(destination: UserRoomsView()) {
-                        VStack (alignment: .center, spacing: 4){
-                            AsyncModelThumbnailView(fileRef: fileRef)
-                            
-                            VStack(alignment: .leading, spacing: 4) {
-                                Text(fileRef.name)
-                                    .font(
-                                        Font.custom("Cambay-Regular", size: 12)
-                                    )
-                                    .foregroundColor(Color("AccentColor"))
+            ScrollView {
+                LazyVGrid(columns: [GridItem(.flexible(), spacing: 16),
+                                    GridItem(.flexible(), spacing: 16)], spacing: 16) {
+                    ForEach(usdzFiles, id: \.self) { fileRef in
+                        NavigationLink(destination: RoomLoaderView(fileRef: fileRef)) {
+                            VStack (alignment: .center, spacing: 4){
+                                AsyncModelThumbnailView(fileRef: fileRef)
+                                
+                                VStack(alignment: .leading, spacing: 4) {
+                                    Text(fileRef.name)
+                                        .font(
+                                            Font.custom("Cambay-Regular", size: 12)
+                                        )
+                                        .foregroundColor(Color("AccentColor"))
+                                }
                             }
+                            .padding([.horizontal, .top])
+                            .background(Color.white)
+                            .cornerRadius(8)
+                            .shadow(color: Color.black.opacity(0.1), radius: 5, x: 0, y: 2)
                         }
-                        .padding([.horizontal, .top])
-                        .background(Color.white)
-                        .cornerRadius(8)
-                        .shadow(color: Color.black.opacity(0.1), radius: 5, x: 0, y: 2)
                     }
                 }
+                Spacer()
             }
-            Spacer()
         }
         .onAppear(perform: {
             userManager.fetchRooms { usdzFiles in
